requirements.txt
flask
flask-mail
python-dotenv
flask_socketio
datetime
cryptography
pytest
pytest-flask
